+++ src/System.Device.Gpio/Interop/Windows/WinRT/Interop.WaitForCompletion.cs
